poubelle-intelligente/requirements.txt
Flask
tensorflow
numpy
pandas
matplotlib
scikit-learn
streamlit
Pillow
opencv-python
seaborn